Separates out the self referencing entity trick to different files
