+++ Salesforce/Synchronize/EventModel/Target.php
@@ -21,6 +21,8 @@ class Target
     private $QBImpossible = false;
     private $QBImpossibleReason = '';
 
+    public $temporaryTables = [];
+
     /**
      * @return LocationQuery[]
      */
